smart statistics update and cleanup
