+++ multimodal-ui/package.json
@@ -25,6 +25,7 @@
     "@angular/platform-browser-dynamic": "^19.1.0",
     "@angular/router": "^19.1.0",
     "@bluehalo/ngx-leaflet": "^19.0.0",
+    "@googlemaps/polyline-codec": "^1.0.28",
     "@types/stats.js": "^0.17.3",
     "d3-color": "^3.1.0",
     "d3-interpolate": "^3.0.1",
@@ -38,7 +39,6 @@
     "node-random-name": "^1.0.1",
     "pixi-filters": "^5.3.0",
     "pixi.js": "^7.4.2",
-    "polyline": "^0.2.0",
     "prettier": "^3.4.2",
     "prettier-eslint": "^16.3.0",
     "rxjs": "~7.8.0",
@@ -56,7 +56,6 @@
     "@types/leaflet": "^1.9.16",
     "@types/node": "^22.14.1",
     "@types/node-random-name": "^1.0.2",
-    "@types/polyline": "^0.1.32",
     "angular-eslint": "19.0.2",
     "eslint": "^9.16.0",
     "jasmine-core": "~5.5.0",
