--- multimodal-ui/src/app/components/recursive-statistic/recursive-statistic.component.html
@@ -1,25 +1,0 @@
-<mat-accordion multi="true">
-  @for (key of keys(); track key) {
-    @if (typeof recursiveDict[key] !== "object") {
-      <div class="flex-column gap-0-5-rem indent">
-        <span style="white-space: normal">
-          {{ capitalize(key) }}:
-          <span style="white-space: nowrap">
-            {{ formatEntry(recursiveDict[key]) }}
-          </span>
-        </span>
-      </div>
-    } @else {
-      <mat-expansion-panel>
-        <mat-expansion-panel-header>
-          <mat-panel-title>
-            {{ capitalize(key) }}
-          </mat-panel-title>
-        </mat-expansion-panel-header>
-        <div class="flex-column gap-0-5-rem">
-          <app-recursive-statistic [recursiveDict]="recursiveDict[key]" />
-        </div>
-      </mat-expansion-panel>
-    }
-  }
-</mat-accordion>
