+++ multimodal-ui/src/app/services/data.service.ts
@@ -5,12 +5,10 @@ import {
   signal,
   WritableSignal,
 } from '@angular/core';
-import { SIMULATION_SAVE_FILE_SEPARATOR } from '../../environments/environment';
 import {
+  extractSimulations,
   Simulation,
-  SIMULATION_STATUSES,
-  SIMULATION_STATUSES_ORDER,
-  SimulationStatus,
+  sortSimulations,
 } from '../interfaces/simulation.model';
 import { CommunicationService } from './communication.service';
 
@@ -39,9 +37,7 @@ export class DataService {
 
   // MARK: Getters
   get simulationsSignal(): Signal<Simulation[]> {
-    return computed(() =>
-      this._simulationsSignal().sort(this.sortSimulations.bind(this)),
-    );
+    return computed(() => this._simulationsSignal().sort(sortSimulations));
   }
 
   get availableSimulationDataSignal(): Signal<string[]> {
@@ -54,22 +50,28 @@ export class DataService {
   }
 
   private listen() {
-    this.communicationService.on('simulations', (simulations) => {
-      this._simulationsSignal.set(
-        this.extractSimulations(simulations as Simulation[]),
-      );
+    this.communicationService.on('simulations', (data) => {
+      const simulations = extractSimulations(data);
+
+      if (simulations === null) {
+        console.error('Invalid simulations data', data);
+        return;
+      }
+
+      this._simulationsSignal.set(simulations);
     });
 
-    this.communicationService.on('available-data', (availableData) => {
-      this._availableSimulationDataSignal.set(
-        (availableData as string[]).sort(),
-      );
-    });
+    this.communicationService.on('available-data', (data) => {
+      if (
+        !Array.isArray(data) ||
+        !data.every((item) => typeof item === 'string')
+      ) {
+        console.error('Invalid available data', data);
+        return;
+      }
 
-    // TODO Uncomment for debugging
-    // this.communicationService.on('log', (data) => {
-    //   console.debug(data);
-    // });
+      this._availableSimulationDataSignal.set(data);
+    });
   }
 
   private query() {
@@ -79,185 +81,5 @@ export class DataService {
 
   private querySimulations() {
     this.communicationService.emit('get-simulations');
-  }
-
-  // MARK: Extraction
-  /**
-   * Validate and extract simulation from the raw data.
-   */
-  private extractSimulations(data: Simulation[]): Simulation[] {
-    return data
-      .map((rawSimulation) => {
-        // TODO Uncomment for debugging
-        // console.debug('Extracting simulation: ', rawSimulation);
-
-        if (!rawSimulation) {
-          console.error('Invalid simulation data: ', rawSimulation);
-          return null;
-        }
-
-        const id = rawSimulation.id;
-        if (!id) {
-          console.error('Simulation ID not found: ', id);
-          return null;
-        }
-
-        const status: SimulationStatus = rawSimulation.status;
-        if (!status) {
-          console.error('Simulation status not found: ', status);
-          return null;
-        }
-        if (!SIMULATION_STATUSES.includes(status)) {
-          console.error('Simulation status not recognized: ', status);
-          return null;
-        }
-
-        if (status === 'corrupted') {
-          return {
-            id,
-            name: id.split(SIMULATION_SAVE_FILE_SEPARATOR)[1] ?? 'unknown',
-            data: 'unknown',
-            status,
-            startTime:
-              this.extractStartTime(
-                id.split(SIMULATION_SAVE_FILE_SEPARATOR)[0] ?? '',
-              ) ?? new Date(0),
-            simulationStartTime: null,
-            simulationEndTime: null,
-            simulationTime: null,
-            simulationEstimatedEndTime: null,
-            lastUpdateIndex: null,
-            completion: 1,
-            configuration: {
-              maxDuration: null,
-            },
-            polylinesVersion: -1,
-            size: null,
-          };
-        }
-
-        const name: string = rawSimulation.name;
-        if (!name) {
-          console.error('Simulation name not found: ', name);
-          return null;
-        }
-
-        const data: string = rawSimulation.data;
-        if (!data) {
-          console.error('Simulation data not found: ', data);
-          return null;
-        }
-
-        const rawStartTime: string =
-          rawSimulation.startTime as unknown as string;
-
-        const startTime: Date | null = this.extractStartTime(rawStartTime);
-        if (startTime === null) {
-          console.error('Invalid simulation start time: ', rawStartTime);
-          return null;
-        }
-
-        const simulationStartTime: number | null =
-          rawSimulation.simulationStartTime ?? null;
-
-        const simulationEndTime: number | null =
-          rawSimulation.simulationEndTime ?? null;
-
-        const simulationTime = rawSimulation.simulationTime ?? null;
-
-        const simulationEstimatedEndTime =
-          rawSimulation.simulationEstimatedEndTime ?? null;
-
-        const lastUpdateIndex = rawSimulation.lastUpdateIndex ?? null;
-
-        let completion = 1;
-        if (
-          simulationStartTime !== null &&
-          simulationTime !== null &&
-          simulationEstimatedEndTime !== null
-        ) {
-          completion =
-            (simulationTime - simulationStartTime) /
-            (simulationEstimatedEndTime - simulationStartTime);
-        }
-
-        const maxDuration = rawSimulation.configuration?.maxDuration ?? null;
-
-        const polylinesVersion = rawSimulation.polylinesVersion ?? -1;
-
-        const size = rawSimulation.size ?? null;
-
-        return {
-          id,
-          name,
-          data,
-          status,
-          startTime,
-          simulationStartTime,
-          simulationEndTime,
-          simulationTime,
-          simulationEstimatedEndTime,
-          lastUpdateIndex,
-          completion,
-          configuration: {
-            maxDuration,
-          },
-          polylinesVersion,
-          size,
-        };
-      })
-      .filter((simulation) => !!simulation);
-  }
-
-  private sortSimulations(a: Simulation, b: Simulation): number {
-    // First compare the orders
-    const aOrder = SIMULATION_STATUSES_ORDER[a.status];
-    const bOrder = SIMULATION_STATUSES_ORDER[b.status];
-
-    if (aOrder < bOrder) {
-      return -1;
-    }
-    if (aOrder > bOrder) {
-      return 1;
-    }
-
-    // If the orders are the same, compare the start times
-    if (a.startTime < b.startTime) {
-      return 1;
-    }
-    return -1;
-  }
-
-  private extractStartTime(rawStartTime: string): Date | null {
-    if (!rawStartTime) {
-      console.error(`Simulation start time not found: ${rawStartTime}`);
-      return null;
-    }
-
-    // Verify the format of the start time
-    if (!/^\d{8}-\d{9}$/.test(rawStartTime)) {
-      console.error(`Invalid format for start time: ${rawStartTime}`);
-      return null;
-    }
-
-    const year: number = parseInt(rawStartTime.slice(0, 4));
-    const month: number = parseInt(rawStartTime.slice(4, 6)) - 1;
-    const day: number = parseInt(rawStartTime.slice(6, 8));
-    const hours: number = parseInt(rawStartTime.slice(9, 11));
-    const minutes: number = parseInt(rawStartTime.slice(11, 13));
-    const seconds: number = parseInt(rawStartTime.slice(13, 15));
-    const milliseconds: number = parseInt(rawStartTime.slice(16, 19));
-
-    const startTime: Date = new Date(
-      year,
-      month,
-      day,
-      hours,
-      minutes,
-      seconds,
-      milliseconds,
-    );
-
-    return startTime;
   }
 }
