+++ python/multimodalsim_viewer/server/data_manager.py
@@ -4,14 +4,12 @@ import os
 from filelock import FileLock
 
 from multimodalsim_viewer.common.utils import (
+    INPUT_DATA_DIRECTORY_PATH,
     SIMULATION_SAVE_FILE_SEPARATOR,
-    get_data_directory_path,
 )
-from multimodalsim_viewer.server.model import (
-    SimulationInformation,
-    Update,
-    VisualizedEnvironment,
-)
+from multimodalsim_viewer.models.environment import VisualizedEnvironment
+from multimodalsim_viewer.models.simulation_information import SimulationInformation
+from multimodalsim_viewer.models.update import Update
 
 
 # MARK: Data Manager
@@ -58,7 +56,8 @@ class SimulationVisualizationDataManager:  # pylint: disable=too-many-public-met
     @staticmethod
     def get_saved_simulations_directory_path() -> str:
         directory_path = os.path.join(
-            get_data_directory_path(), SimulationVisualizationDataManager.__SAVED_SIMULATIONS_DIRECTORY_NAME
+            SimulationVisualizationDataManager.get_data_directory_path(),
+            SimulationVisualizationDataManager.__SAVED_SIMULATIONS_DIRECTORY_NAME,
         )
 
         if not os.path.exists(directory_path):
@@ -555,3 +554,48 @@ class SimulationVisualizationDataManager:  # pylint: disable=too-many-public-met
                     updates[update_index] = current_state_updates
 
         return states, updates
+
+    # MARK: +- Simulation Data
+    @staticmethod
+    def get_data_directory_path() -> str:
+        current_file_path = os.path.abspath(__file__)
+        current_file_dir = os.path.dirname(current_file_path)
+        data_directory_path = os.path.join(current_file_dir, "..", "data")
+
+        if not os.path.exists(data_directory_path):
+            os.makedirs(data_directory_path)
+
+        return data_directory_path
+
+    @staticmethod
+    def get_saved_logs_directory_path() -> str:
+        data_directory_path = SimulationVisualizationDataManager.get_data_directory_path()
+        saved_logs_directory_path = os.path.join(data_directory_path, "saved_logs")
+
+        if not os.path.exists(saved_logs_directory_path):
+            os.makedirs(saved_logs_directory_path)
+
+        return saved_logs_directory_path
+
+    @staticmethod
+    def get_input_data_directory_path(data: str | None = None) -> str:
+        input_data_directory = INPUT_DATA_DIRECTORY_PATH
+
+        if data is not None:
+            input_data_directory = os.path.join(input_data_directory, data)
+
+        return input_data_directory
+
+    @staticmethod
+    def get_available_data():
+        input_data_directory = SimulationVisualizationDataManager.get_input_data_directory_path()
+
+        if not os.path.exists(input_data_directory):
+            return []
+
+        # List all directories in the input data directory
+        return [
+            name
+            for name in os.listdir(input_data_directory)
+            if os.path.isdir(os.path.join(input_data_directory, name)) and not name.startswith(".")
+        ]
