+++ python/multimodalsim_viewer/server/log_manager.py
@@ -1,8 +1,8 @@
-from multimodalsim_viewer.common.utils import get_saved_logs_directory_path
+from multimodalsim_viewer.server.data_manager import SimulationVisualizationDataManager
 
 
 def register_log(simulation_id, message):
-    saved_logs_directory_path = get_saved_logs_directory_path()
+    saved_logs_directory_path = SimulationVisualizationDataManager.get_saved_logs_directory_path()
     file_name = f"{simulation_id}.txt"
     file_path = f"{saved_logs_directory_path}/{file_name}"
